Update code to improve synchronization and refine raid details
The main changes in this commit are to improve account sync for character and raid info where the RaidEncounter method is deprecated and replaced by the RaidDetail method. Also multiple enhancements are made to the RaidDetailEntity. These changes include adding new fields for improved detail level and implementing logic to calculate existing raid details ID's to avoid duplicate entries. Furthermore, this commit introduces the Difficulty Converter and enum for handling raid difficulty levels. Tests are updated to reflect these changes as well.